LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/redirect-to-about.window-expected.txt

CONSOLE MESSAGE: Not allowed to redirect to about:blank due to its scheme
CONSOLE MESSAGE: Not allowed to redirect to about:blank due to its scheme
CONSOLE MESSAGE: Not allowed to redirect to about:srcdoc due to its scheme
CONSOLE MESSAGE: Not allowed to redirect to about:srcdoc due to its scheme
CONSOLE MESSAGE: Not allowed to redirect to about:nonstandard due to its scheme
CONSOLE MESSAGE: Not allowed to redirect to about:nonstandard due to its scheme

PASS An iframe with src set to a redirect to about:blank
PASS An iframe that is navigated to a redirect to about:blank
PASS An iframe with src set to a redirect to about:srcdoc
PASS An iframe that is navigated to a redirect to about:srcdoc
PASS An iframe with src set to a redirect to about:nonstandard
PASS An iframe that is navigated to a redirect to about:nonstandard
